+++ packages/pluggableWidgets/maps-web/src/Maps.tsx
@@ -36,8 +36,8 @@ export default function Maps(props: MapsContainerProps): ReactNode {
             heightUnit={props.heightUnit}
             locations={locations}
             mapsToken={props.apiKeyExp?.value ?? props.apiKey}
+            mapId={props.googleMapId}
             mapProvider={props.mapProvider}
-            mapStyles={props.mapStyles}
             mapTypeControl={props.mapTypeControl}
             optionDrag={props.optionDrag}
             optionScroll={props.optionScroll}