+++ packages/pluggableWidgets/maps-web/src/components/__tests__/GoogleMap.spec.ts
@@ -1,21 +1,7 @@
 import { shallow, ShallowWrapper } from "enzyme";
 import { createElement } from "react";
-import { GoogleMap, GoogleMapsProps } from "../GoogleMap";
+import { GoogleMapContainer, GoogleMapsProps } from "../GoogleMap";
 import { initialize } from "@googlemaps/jest-mocks";
-import { useLoadScript } from "@react-google-maps/api";
-
-jest.mock("@react-google-maps/api", () => {
-    const original = jest.requireActual("@react-google-maps/api");
-    return {
-        ...original,
-        useLoadScript: jest.fn()
-    };
-});
-
-function mockUseLoadScriptHookWithReturn(returnValue: Partial<ReturnType<typeof useLoadScript>>): void {
-    // @ts-expect-error `mockImplementation` is present cuz of the mock, but TS doesn't know.
-    useLoadScript.mockImplementation(() => returnValue);
-}
 
 describe("Google maps", () => {
     const defaultProps: GoogleMapsProps = {
@@ -26,8 +12,8 @@ describe("Google maps", () => {
         height: 75,
         heightUnit: "pixels",
         locations: [],
-        mapStyles: "",
         mapsToken: "",
+        mapId: "DEMO_MAP_ID",
         mapTypeControl: false,
         optionDrag: true,
         optionScroll: true,
@@ -43,7 +29,6 @@ describe("Google maps", () => {
 
     beforeEach(() => {
         initialize();
-        mockUseLoadScriptHookWithReturn({ isLoaded: true, loadError: undefined });
     });
 
     afterEach(() => {
@@ -51,7 +36,7 @@ describe("Google maps", () => {
     });
 
     const renderGoogleMap = (props: GoogleMapsProps): ShallowWrapper<GoogleMapsProps, any> =>
-        shallow(createElement(GoogleMap, props));
+        shallow(createElement(GoogleMapContainer, props));
 
     it("renders a map with right structure", () => {
         const googleMaps = renderGoogleMap(defaultProps);