+++ packages/pluggableWidgets/maps-web/typings/MapsProps.d.ts
@@ -96,7 +96,7 @@ export interface MapsContainerProps {
     height: number;
     zoom: ZoomEnum;
     mapProvider: MapProviderEnum;
-    mapStyles: string;
+    googleMapId: string;
 }
 
 export interface MapsPreviewProps {
@@ -130,5 +130,5 @@ export interface MapsPreviewProps {
     height: number | null;
     zoom: ZoomEnum;
     mapProvider: MapProviderEnum;
-    mapStyles: string;
+    googleMapId: string;
 }